@@ -4,8 +4,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import crawlerapi.entity.User;
 import crawlerapi.security.JWTUtil;
 import crawlerapi.security.PBKDF2Encoder;
 import crawlerapi.security.model.AuthRequest;
@@ -15,6 +17,7 @@ import lombok.RequiredArgsConstructor;
 import reactor.core.publisher.Mono;
 
 @RestController
+@RequestMapping("/crawler-api")
 @RequiredArgsConstructor
 public class AuthenticationController {
 
@@ -26,9 +29,9 @@ public class AuthenticationController {
 
     @PostMapping("/login")
     public Mono<ResponseEntity<?>> login(@RequestBody AuthRequest ar) {
-        return userRepository.findByUsername(ar.getUsername()).map(userDetails -> {
-            if (passwordEncoder.encode(ar.getPassword()).equals(userDetails.getPassword())) {
-                return ResponseEntity.ok(new AuthResponse(jwtUtil.generateToken(userDetails)));
+        return Mono.just(userRepository.findByUsername(ar.getUsername())).map(userDetails -> {
+            if (passwordEncoder.encode(ar.getPassword()).equals(userDetails.orElseGet(User::new).getPassword())) {
+                return ResponseEntity.ok(new AuthResponse(jwtUtil.generateToken(userDetails.orElseGet(User::new))));
             } else {
                 return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
             }
